requirements.txt
torch
transformers
sentencepiece
accelerate
bitsandbytes
python-dotenv